This fixes #147
The dashboard now acts as startscreen of the app. This means that if you
click the back button on the dashboard at any time you will exit the
app. To achieve this behavior, the backstack is emptied if the dashboard
is selected. Also, it is checked now before selection of a view, whether
the view is already shown on the screen.