Localized "loading..." text.
The "loading..." text will now also be shown for the fuel price
calculation. If the necessary fuel consumption value is available and
the app is asking the server for the fuel price the "loading..." text
will be replaced by "calculating..." and this will be replaced by the
actual fuel price.